Refactor: Update Gradle tasks in GitHub workflows
This commit updates the Gradle tasks used in the GitHub workflows to reflect changes in the project's build configuration. Specifically, it updates tasks related to screenshot tests, builds, lint checks, badging, and connected tests to use the "DemofossReliant" flavor instead of "FossReliant". It also adjusts the assemble tasks to use "ProdGooglePlayRelease" and "ProdfossReliantRelease" for production builds and updates the corresponding APK output directories.